Update configuration.nix and packages.nix
- Added XDG Portal configuration
- Added xdg-utils, pavucontrol, and easyeffects to packages.nix
- Updated update.sh script to handle laptop or desktop argument

+++ update.sh
@@ -1,13 +1,23 @@
-# Remove content in /etc/nixos
-rm -rf /etc/nixos/*
-echo "Removed content in /etc/nixos"
+if [ "$1" != "laptop" -a "$1" != "desktop" ]
+then
+	echo "Argument attendu : 'laptop' ou 'desktop'"
+	exit 1
+else
+	# Remove content in /etc/nixos
+	rm -rf /etc/nixos/*
+	echo "Removed content in /etc/nixos"
 
-# Copy content from ~/Documents/nix-configuration to /etc/nixos
-# cp -r /home/laptop/Documents/nix-configuration/* /etc/nixos/
-cp -r /home/desktop/Documents/nix-configuration/* /etc/nixos/
-# Remove update.sh from /etc/nixos
-rm /etc/nixos/update.sh
-echo "Copied content from ~/Documents/nix-configuration to /etc/nixos"
+	# Copy content from ~/Documents/nix-configuration to /etc/nixos
+	cp -r /home/$1/Documents/nix-configuration/* /etc/nixos/
+	echo "NixOS configuration copied to /etc/nixos"
 
-# Update NixOS
-cd /etc/nixos && nixos-rebuild switch --flake .#desktop --show-trace
+	# Remove update.sh from /etc/nixos
+	rm /etc/nixos/update.sh
+	echo "Removed update.sh from /etc/nixos"
+
+	# Update NixOS
+	cd /etc/nixos && nixos-generate-config && nixos-rebuild switch --flake .#$1 --show-trace
+fi
+
+
+
